Now saves "about" information in export
We're scanning "about.json" for relevant metadata, and making that
available in a new 'about' folder in plain text, HTML, and original
JSON formats; when pandoc is available, we're also saving it in
Markdown.

This metadata includes the list name, description, website, email,
trailer, location, language, country, and category. List permissions
can be funky and complicated, so we pretty-print that into a single
English string.

The original about.json data can have slightly funky HTML, so we do
some cleanup before passing it along.

We also save the original "about.json" file. This may be the only case
where we're saving original API output, instead of something cleaned
up and processed. We're doing this because it's just the one little
file, and there could be situations where someone wants to look at the
original metadata fields.